pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-safe_station"
version = "0.1.0"
description = "Safe Station environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi",
    "uvicorn",
    "pydantic",
    "httpx",
    "python-multipart",
    "websockets",
    "psutil",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server", "."]
package-dir = { "server" = "server", "" = "." }